rateyourprof_api/requirements.txt
alembic==0.9.6
aniso8601==2.0.0
asn1crypto==0.24.0
attrs==17.4.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cryptography==2.1.4
Flask==0.12.2
Flask-Cors==3.0.3
Flask-JWT-Extended==3.6.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
pycparser==2.18
PyJWT==1.5.3
pytest==3.4.0
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
urllib3==1.22
uuid==1.30
Werkzeug==0.14.1
